Changes by Komal

+++ src/main/java/com/mahait/gov/in/service/OrganizationInstInfoService.java
@@ -18,7 +18,7 @@ public interface OrganizationInstInfoService {
 
 	public List<Object[]> getBankBranch(String valueOf);
 
-	public int SaveorgInstituteInfo(@Valid OrgDdoMstModel orgDdoMstModel);
+	public Long SaveorgInstituteInfo(@Valid OrgDdoMstModel orgDdoMstModel);
 
 	public int updateorgInstituteInfo(OrgDdoMstModel orgDdoMstModel);
